requirements.txt
certifi==2018.4.16
chardet==3.0.4
future==0.16.0
idna==2.6
mqttclient==0.0.2
onoff==1.0.1
paho-mqtt==1.3.1
PySocks==1.6.8
pyTelegramBotAPI==3.6.3
requests==2.18.4
six==1.11.0
urllib3==1.22
wheezy.routing==0.1.157
